WIP on master: 512c055 Fixed loading of png's in Drag and Drop on Windows.
Former-commit-id: 77b6174d8dbc1757e5634986b2b656dc77a28a7e

+++ mrViewer/src/gui/mrvPreferences.cpp
@@ -1071,7 +1071,7 @@ static const char* kCLocale = "C";
 	view->toggle_presentation();
       }
 
-    GLLut3d::NUM_STOPS = uiPrefs->uiPrefsNumStops->value();
+    GLLut3d::NUM_STOPS = (unsigned) uiPrefs->uiPrefsNumStops->value();
     
     int num = (int)main->uiPrefs->uiPrefsOpenEXRThreadCount->value();
     Imf::setGlobalThreadCount( num );
